Move HTTP event client to new event-http module
